+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Changes in 0.10.1 - DEV
 - Remove test dependencies (nose and rednose) from install dependencies list. #1079
 - Recursively build query when using elemMatch operator. #1130
 - Fix instance back references for lists of embedded documents. #1131
+- Allow shard key to point to a field in an embedded document. #551
 
 Changes in 0.10.0
 =================